events: Refactor pacts test code
